Implemented changes per review, cleaned up test
